+++ js/utils.js
@@ -1,0 +1,4 @@
+// Генератор случайных значений
+const randomCalculate = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
+
+export {randomCalculate};
